refactor(test): extract recovery tests from proxy-handler.test.ts (#626)
- New proxy-handler-recovery.test.ts (387 lines): previous_response_not_found,
  unanswered function_call, WebSocket fallback, ban-state descriptive errors
- New helpers/proxy-handler-fixtures.ts: shared createMockAccountPool,
  createDefaultRequest, buildTestApp utilities
- Original proxy-handler.test.ts trimmed from 1523 → 1038 lines

All 34 tests pass across both files.
