+++ tests/integration/proxy-handler.test.ts
@@ -1035,294 +1035,4 @@ describe("proxy-handler integration", () => {
     expect(accountPool.acquire).toHaveBeenCalledTimes(1);
   });
 
-  // 17. previous_response_not_found: should strip previous_response_id and retry
-  // Reproduces the bug: when upstream returns previous_response_not_found
-  // (because the response was created on a different account or is unknown to
-  // this account), the proxy currently passes the error straight through with
-  // no recovery. Desired behavior: strip previous_response_id and replay.
-  it("recovers from previous_response_not_found by stripping ID and retrying", async () => {
-    const notFoundBody = JSON.stringify({
-      error: {
-        type: "invalid_request_error",
-        code: "previous_response_not_found",
-        message: "Previous response with id 'resp_0e2e6e7917486cfd0069eec8532d988194a3da6379c70abe68' not found.",
-      },
-    });
-
-    let createCount = 0;
-    const seenPrevIds: Array<string | undefined> = [];
-    mockCreateResponse = () => {
-      // The mock-CodexApi instance receives the *current* codexRequest by
-      // reference. We capture the previous_response_id at the moment of call.
-      createCount++;
-      if (createCount === 1) return Promise.reject(new CodexApiError(400, notFoundBody));
-      return Promise.resolve(new Response("data: {}\n\n"));
-    };
-
-    // Wrap createResponse to capture the previous_response_id seen on each call.
-    // (Re-mock CodexApi inline so we can observe the request mutation.)
-    const accountPool = createMockAccountPool();
-    const fmt = createMockFormatAdapter();
-    const req: ProxyRequest = {
-      ...createDefaultRequest(),
-      codexRequest: {
-        ...createDefaultRequest().codexRequest,
-        previous_response_id: "resp_0e2e6e7917486cfd0069eec8532d988194a3da6379c70abe68",
-      },
-    };
-
-    // Spy: snapshot previous_response_id before the upstream is called.
-    const origMock = mockCreateResponse;
-    mockCreateResponse = () => {
-      seenPrevIds.push(req.codexRequest.previous_response_id);
-      return origMock!();
-    };
-
-    const { app } = buildTestApp({ accountPool, fmt, req });
-    const res = await app.request("/test", { method: "POST" });
-
-    // Desired: proxy retries after stripping previous_response_id, returns 200
-    expect(res.status).toBe(200);
-    // Desired: 2 upstream calls — first with the stale ID, second without it
-    expect(createCount).toBe(2);
-    expect(seenPrevIds[0]).toBe("resp_0e2e6e7917486cfd0069eec8532d988194a3da6379c70abe68");
-    expect(seenPrevIds[1]).toBeUndefined();
-  });
-
-  it("replays full original input after implicit previous-response WebSocket failure", async () => {
-    const req: ProxyRequest = {
-      ...createDefaultRequest(),
-      codexRequest: {
-        ...createDefaultRequest().codexRequest,
-        prompt_cache_key: "thread-implicit-ws",
-        input: [
-          { role: "user", content: "first" },
-          { role: "assistant", content: "ok" },
-          { role: "user", content: "continue" },
-        ],
-        turnState: "turn-original",
-        useWebSocket: false,
-      },
-    };
-    const affinityMap = getSessionAffinityMap();
-    const promptCacheIdentity = resolvePromptCacheIdentity(req.codexRequest, req.clientConversationId);
-    const variantHash = computeVariantHash(
-      req.codexRequest.instructions,
-      req.codexRequest.tools,
-      buildVariantIdentity(req.codexRequest, promptCacheIdentity),
-    );
-    affinityMap.record(
-      "resp_implicit_ws",
-      "e1",
-      "thread-implicit-ws",
-      "turn-implicit",
-      "You are helpful",
-      undefined,
-      undefined,
-      variantHash,
-    );
-
-    const seenRequests: Array<{
-      input: CodexResponsesRequest["input"];
-      previousResponseId: string | undefined;
-      turnState: string | undefined;
-      useWebSocket: boolean | undefined;
-    }> = [];
-    let createCount = 0;
-    mockCreateResponse = (request) => {
-      createCount++;
-      seenRequests.push({
-        input: [...request.input],
-        previousResponseId: request.previous_response_id,
-        turnState: request.turnState,
-        useWebSocket: request.useWebSocket,
-      });
-      if (createCount === 1) {
-        return Promise.reject(new PreviousResponseWebSocketError("ws down"));
-      }
-      return Promise.resolve(new Response("data: {}\n\n"));
-    };
-
-    const accountPool = createMockAccountPool();
-    const fmt = createMockFormatAdapter();
-    const { app } = buildTestApp({ accountPool, fmt, req });
-
-    const res = await app.request("/test", { method: "POST" });
-    expect(res.status).toBe(200);
-
-    expect(seenRequests).toEqual([
-      {
-        input: [{ role: "user", content: "continue" }],
-        previousResponseId: "resp_implicit_ws",
-        turnState: "turn-implicit",
-        useWebSocket: true,
-      },
-      {
-        input: [
-          { role: "user", content: "first" },
-          { role: "assistant", content: "ok" },
-          { role: "user", content: "continue" },
-        ],
-        previousResponseId: undefined,
-        turnState: "turn-original",
-        useWebSocket: false,
-      },
-    ]);
-    expect(accountPool.acquire).toHaveBeenCalledTimes(1);
-    expect(accountPool.release).toHaveBeenCalledWith("e1", {
-      input_tokens: 10,
-      output_tokens: 20,
-    });
-  });
-
-  it("recovers when collectTranslator raises previous_response_not_found", async () => {
-    const notFoundBody = JSON.stringify({
-      error: {
-        type: "invalid_request_error",
-        code: "previous_response_not_found",
-        message: "Previous response with id 'resp_collect_stale' not found.",
-      },
-    });
-    const req: ProxyRequest = {
-      ...createDefaultRequest(),
-      codexRequest: {
-        ...createDefaultRequest().codexRequest,
-        previous_response_id: "resp_collect_stale",
-      },
-    };
-    let createCount = 0;
-    const seenPrevIds: Array<string | undefined> = [];
-    mockCreateResponse = () => {
-      createCount++;
-      seenPrevIds.push(req.codexRequest.previous_response_id);
-      return Promise.resolve(new Response("data: {}\n\n"));
-    };
-
-    let collectCount = 0;
-    const fmt = createMockFormatAdapter({
-      collectTranslator: vi.fn(async () => {
-        collectCount++;
-        if (collectCount === 1) {
-          throw new CodexApiError(400, notFoundBody);
-        }
-        return {
-          response: { id: "resp_after_collect_retry", choices: [] },
-          usage: { input_tokens: 7, output_tokens: 3 },
-          responseId: "resp_after_collect_retry",
-        };
-      }),
-    });
-    const accountPool = createMockAccountPool();
-    const { app } = buildTestApp({ accountPool, fmt, req });
-
-    const res = await app.request("/test", { method: "POST" });
-    expect(res.status).toBe(200);
-
-    expect(createCount).toBe(2);
-    expect(collectCount).toBe(2);
-    expect(seenPrevIds[0]).toBe("resp_collect_stale");
-    expect(seenPrevIds[1]).toBeUndefined();
-    expect(accountPool.release).toHaveBeenCalledTimes(1);
-    expect(accountPool.release).toHaveBeenCalledWith("e1", {
-      input_tokens: 7,
-      output_tokens: 3,
-    });
-  });
-
-  // 17b. previous_response_not_found loop guard — only retry once
-  it("does not loop forever when previous_response_not_found persists after strip", async () => {
-    const notFoundBody = JSON.stringify({
-      error: { type: "invalid_request_error", code: "previous_response_not_found",
-        message: "Previous response with id 'resp_xxx' not found." },
-    });
-    let createCount = 0;
-    mockCreateResponse = () => {
-      createCount++;
-      return Promise.reject(new CodexApiError(400, notFoundBody));
-    };
-
-    const accountPool = createMockAccountPool();
-    const fmt = createMockFormatAdapter();
-    const req: ProxyRequest = {
-      ...createDefaultRequest(),
-      codexRequest: {
-        ...createDefaultRequest().codexRequest,
-        previous_response_id: "resp_xxx",
-      },
-    };
-    const { app } = buildTestApp({ accountPool, fmt, req });
-
-    const res = await app.request("/test", { method: "POST" });
-    // After the strip+retry attempt also fails, fall through to generic error
-    expect(res.status).toBe(400);
-    // Exactly 2 upstream calls — strip-retry happens once, no further retries
-    expect(createCount).toBe(2);
-  });
-
-  // 17c. unanswered function_call: upstream "No tool output found for function
-  // call call_X" means a stored function_call from the previous response was
-  // not answered. Recovery: strip previous_response_id, retry once on the same
-  // account (full input replay covers the missing context).
-  it("recovers from unanswered function_call by stripping ID and retrying", async () => {
-    const unansweredBody = JSON.stringify({
-      error: {
-        type: "invalid_request_error",
-        message: "No tool output found for function call call_8vO7oqvintBWH5bAoAz3vPh5.",
-      },
-    });
-
-    let createCount = 0;
-    const seenPrevIds: Array<string | undefined> = [];
-    const req: ProxyRequest = {
-      ...createDefaultRequest(),
-      codexRequest: {
-        ...createDefaultRequest().codexRequest,
-        previous_response_id: "resp_unanswered_chain",
-      },
-    };
-    mockCreateResponse = () => {
-      seenPrevIds.push(req.codexRequest.previous_response_id);
-      createCount++;
-      if (createCount === 1) return Promise.reject(new CodexApiError(400, unansweredBody));
-      return Promise.resolve(new Response("data: {}\n\n"));
-    };
-
-    const accountPool = createMockAccountPool();
-    const fmt = createMockFormatAdapter();
-    const { app } = buildTestApp({ accountPool, fmt, req });
-    const res = await app.request("/test", { method: "POST" });
-
-    expect(res.status).toBe(200);
-    expect(createCount).toBe(2);
-    expect(seenPrevIds[0]).toBe("resp_unanswered_chain");
-    expect(seenPrevIds[1]).toBeUndefined();
-  });
-
-  // 18. 403 ban with mixed pool states → descriptive error
-  it("returns descriptive error when banned and remaining accounts disabled/expired", async () => {
-    mockCreateResponse = () =>
-      Promise.reject(new CodexApiError(403, '{"detail": "Account suspended"}'));
-
-    const accountPool = createMockAccountPool({
-      acquire: vi.fn()
-        .mockReturnValueOnce({ entryId: "e1", token: "tok", accountId: "acc1" }),
-      hasAvailableAccounts: vi.fn(() => false),
-      getPoolSummary: vi.fn(() => ({
-        total: 3, active: 0, expired: 1, quota_exhausted: 0,
-        rate_limited: 0, refreshing: 0, disabled: 1, banned: 1,
-      })),
-    });
-    const fmt = createMockFormatAdapter();
-    const { app } = buildTestApp({ accountPool, fmt });
-
-    const res = await app.request("/test", { method: "POST" });
-    expect(res.status).toBe(403);
-
-    const body = await res.json();
-    expect(body.error).toBe("api_error");
-    expect(body.message).toContain("All accounts exhausted");
-    expect(body.message).toContain("1 expired");
-    expect(body.message).toContain("1 disabled");
-    expect(body.message).toContain("1 banned");
-  });
 });
